pyproject.toml
[project]
name = "glowbridge-agent-ops"
version = "0.1.0"
description = "GlowBridge Call Capture - agent operations backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "anthropic>=0.39.0",
    "supabase>=2.10.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.4",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.20",
    "stripe>=11.0.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.150.0",
    "beautifulsoup4>=4.12.0",
    "tldextract>=5.1.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
